+++ NEWS.md
@@ -1,3 +1,6 @@
+# spant 2.13.0
+* Improved defaults for sim_brain_1h function.
+
 # spant 2.12.0
 * Amplitude scaling functions duplicate water reference amplitudes when
 sensible.